bug_20140306.txt

+++ src/common/tools.cpp
@@ -53,14 +53,14 @@ int createUUID(char uuid[UUID_LEN]){
 	static u8 _rand_init = 0;
 	static char randstate[2048];
 	if(_rand_init == 0){
-		initstate(time(NULL), randstate, 2048);
+		initstate((unsigned)time(NULL), randstate, 2048);
 		setstate(randstate);
 		_rand_init = 1;
 	}
 	
 	u32 r[8];
 	for(int x=0;x<8;x++){
-		r[x] = random();
+		r[x] = (u32)random();
 	}
 	//Make repres
 	//Fix this
@@ -92,7 +92,7 @@ u32 righttrim(u8* buf, u32 len){
 	while(pos>=buf && *pos==' '){
 		pos--;
 	}
-	return (pos-buf)+1;
+	return (u32)((pos-buf)+1);
 }
 
 
@@ -134,7 +134,7 @@ void trimspaces(u8** buf, u32* len){
 	u32 l = *len;
 	//left
 	b = lefttrim(*buf, *len);
-	l = l - (b - *buf);
+	l = (u32)(l - (b - *buf));
 	//right
 	l = righttrim(b, l);
